feat: simplfiied webhook handler
